requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-groq>=0.1.0
langchain-experimental>=0.0.60
langchain-huggingface>=0.0.3
chromadb>=0.5.0
pdfplumber>=0.11.0
python-docx>=1.1.0
sentence-transformers>=3.0.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
pandas>=2.0.0
